Version 0.7: support both HipChat and Campfire credentials in OS X keychain
When running on OS X, clogger will, if no access tokens were found in the
configuration file, look in the keychain for an entry matching either
clogger or its name suffixed with the chat type (e.g. "clogger-hipchat").
If none are found, clogger will prompt the user for the API token and
subsequently store it in the keychain.

This allows the user to store multiple credentials for different service
without the need for an insecure text file.